Added testing for USFM import, organized helpers
