+++ src/js/actions/CSVExportActions.js
@@ -322,4 +322,3 @@ export function getToolFolderNames(projectPath) {
   } catch (e) {
   }
 }
-
