+++ src/js/actions/ProjectSelectionActions.js
@@ -10,6 +10,8 @@ import * as ProjectValidationActions from './ProjectValidationActions';
 // helpers
 import * as ProjectSelectionHelpers from '../helpers/ProjectSelectionHelpers';
 import * as LoadHelpers from '../helpers/LoadHelpers';
+import * as manifestHelpers from '../helpers/manifestHelpers';
+import * as usfmHelpers from '../helpers/usfmHelpers';
 
 
 /**
@@ -26,7 +28,7 @@ export function selectProject(projectPath, projectLink) {
     projectPath = LoadHelpers.saveProjectInHomeFolder(projectPath);
     let manifest, params, targetLanguage;
     /**@type {String} */
-    let USFMFilePath = LoadHelpers.isUSFMProject(projectPath);
+    let USFMFilePath = usfmHelpers.isUSFMProject(projectPath);
     //If present proceed to usfm loading process
     if (USFMFilePath) {
       let usfmProjectObject = ProjectSelectionHelpers.getProjectDetailsFromUSFM(USFMFilePath, projectPath);
@@ -38,7 +40,7 @@ export function selectProject(projectPath, projectLink) {
       //If no usfm file found proceed to load regular loading process
       manifest = ProjectSelectionHelpers.getProjectManifest(projectPath, projectLink, username);
       if (!manifest) dispatch(AlertModalActions.openAlertDialog("No valid manifest found in project"));
-      params = LoadHelpers.getParams(projectPath, manifest);
+      params = manifestHelpers.getParams(projectPath, manifest);
     }
     dispatch(clearLastProject());
     dispatch(loadProjectDetails(projectPath, manifest, params));
@@ -121,7 +123,7 @@ export function displayTools() {
   return ((dispatch, getState) => {
     const { currentSettings } = getState().settingsReducer;
     const { manifest } = getState().projectDetailsReducer;
-    if (LoadHelpers.checkIfValidBetaProject(manifest) || currentSettings.developerMode) {
+    if (manifestHelpers.checkIfValidBetaProject(manifest) || currentSettings.developerMode) {
       dispatch(ToolsMetadataActions.getToolsMetadatas());
       // Go to toolsCards page
       dispatch(BodyUIActions.goToStep(3));
