+++ src/js/helpers/MergeConflictHelpers.js
@@ -9,8 +9,8 @@ const replaceRegex = /(<<<<<<<\s?.*[\s\S]*?>>>>>>>\s?.*)/;
 /**
  * Seaches usfm data with regex and returns all merge conflicts found separated by string.
  * Each element in the array represents a different version of the conflict.
- * The versions are split by pair of two naturally. 
- * 
+ * The versions are split by pair of two naturally.
+ *
  * @example ["1 this is the first version", "1 This is the second version"] - represents one verse conflict
  * @param {string} usfmData - usfm string to be searched for merge conflicts
  * @returns {[string]}
@@ -80,7 +80,7 @@ export function parseMergeConflictVersion(versionText, usfmData) {
  * @param {object} mergeConflictsObject - Object to be parsed that contains the information
  * of the chosen text to merge
  * @param {string} projectSaveLocation - Path of the project
- * @param {object} manifest - Metadata of the project details 
+ * @param {object} manifest - Metadata of the project details
  */
 export function merge(mergeConflictsObject, projectSaveLocation, manifest) {
   try {
@@ -132,7 +132,7 @@ export function createUSFMFromTsProject(projectSaveLocation) {
 }
 
 /**
- * Determines whether or not there is usfm to parse for merge conflicts 
+ * Determines whether or not there is usfm to parse for merge conflicts
  * and if there is return the data it contains
  * @param {string} usfmFilePath - path to the usfm file of the project, note this may not exist
  * @param {string} projectSaveLocation - path to the projects location
@@ -154,4 +154,26 @@ export function checkProjectForMergeConflicts(usfmFilePath, projectSaveLocation)
     } catch (e) { console.warn('Problem getting merge conflicts', e) }
   }
   return usfmData;
+}
+
+/**
+ * This method reads in all the chunks of a project, and determines if there is any merge conflicts.
+ * @param {Array} projectChunks - An array of finished chunks, as defined by the manfest
+ * @param {String} projectPath - The current save location of the project
+ * @returns {Boolean} True if there is any merge conflicts, false if the project does not contain any
+ */
+export function projectHasMergeConflicts(projectPath, bookAbbr) {
+  let currentFolderChapters = fs.readdirSync(Path.join(projectPath, bookAbbr));
+  for (var currentChapterFile of currentFolderChapters) {
+    let currentChapter = Path.parse(currentChapterFile).name;
+    if (!parseInt(currentChapter)) continue;
+    try {
+      let currentChapterObject = fs.readJSONSync(Path.join(projectPath, bookAbbr, currentChapterFile));
+      let fileContents = JSON.stringify(currentChapterObject);
+      if (~fileContents.indexOf('<<<<<<<')) {
+        return true;
+      }
+    } catch (e) { }
+  }
+  return false;
 }
