+++ tests/b.LoadOnline.js
@@ -1,12 +1,15 @@
-const assert = require('chai').assert;
-const loadOnline = require('../src/js/components/LoadOnline.js');
-const path = require('path-extra');
-const fs = require('fs-extra');
+import {describe, it} from 'mocha';
+import {assert} from 'chai';
+import path from 'path-extra';
+import fs from 'fs-extra';
+//helpers
+import * as loadOnline from '../src/js/helpers/LoadOnlineHelpers';
+
 const badSave = path.join(path.homedir(), 'translationCore', 'projects', 'id_-cfksl');
 
 describe('loadOnline.openManifest', function() {
   it('loadOnline.openManifest should return an error if no link is specified', function(done) {
-    loadOnline(null, function(err, savePath, url) {
+    loadOnline.openManifest(null, function(err, savePath, url) {
       assert.isString(err.text);
       assert.isNull(savePath);
       assert.isNull(url);
@@ -16,8 +19,8 @@ describe('loadOnline.openManifest', function() {
   });
   it('loadOnline.openManifest should fail on an invalid link.', function(done){
     fs.removeSync(badSave);
-    this.timeout(500000);
-    loadOnline('https://git.door43.org/ianhoegen123/id_-cfksl.git', function(err, savePath, url) {
+    this.timeout(10000);
+    loadOnline.openManifest('https://git.door43.org/klappy/noprojecthere.git', function(err, savePath, url) {
       assert.isNull(savePath);
       assert.isNull(url);
       assert.isString(err.text);
@@ -26,10 +29,10 @@ describe('loadOnline.openManifest', function() {
   });
 
   it('loadOnline.openManifest should not deny a non .git link.', function(done){
-    this.timeout(50000);
-    var expectedSavePath = path.join(path.homedir(), 'translationCore', 'projects', 'id_-co_text_reg');
-    var expectedURL = 'https://git.door43.org/royalsix/id_-co_text_reg';
-    loadOnline(expectedURL, function(err, savePath, url) {
+    this.timeout(20000);
+    var expectedSavePath = path.join(path.homedir(), 'translationCore', 'projects', 'bhadrawahi_tit');
+    var expectedURL = 'https://git.door43.org/klappy/bhadrawahi_tit';
+    loadOnline.openManifest(expectedURL, function(err, savePath, url) {
       assert.equal(savePath, expectedSavePath);
       assert.equal(url, expectedURL+'.git');
       fs.removeSync(savePath);
@@ -38,10 +41,10 @@ describe('loadOnline.openManifest', function() {
   });
 
   it('loadOnline.openManifest should return the home directory and url', function(done){
-    this.timeout(50000);
-    var expectedSavePath = path.join(path.homedir(), 'translationCore', 'projects', 'id_-co_text_reg');
-    var expectedURL = 'https://git.door43.org/royalsix/id_-co_text_reg.git';
-    loadOnline(expectedURL, function(err, savePath, url) {
+    this.timeout(20000);
+    var expectedSavePath = path.join(path.homedir(), 'translationCore', 'projects', 'bhadrawahi_tit');
+    var expectedURL = 'https://git.door43.org/klappy/bhadrawahi_tit.git';
+    loadOnline.openManifest(expectedURL, function(err, savePath, url) {
       assert.equal(savePath, expectedSavePath);
       assert.equal(url, expectedURL);
       done();
