+++ tests/fixtures/usfm/valid/autographa.usfm
@@ -1,0 +1,75 @@
+\id TIT Bhadrawahi Bible
+\ide UTF-8
+\rem Copyright Information: Creative Commons Attribution-Share Alike 4.0 License
+\h तीतुस
+\toc1 तीतुसेरे नोंवें पौलुसेरी चिठ्ठी
+\toc2 तीतुस
+\mt1 TITUS
+\mt1 तीतुस
+\c 1
+\s अभिवादन
+\p
+\v 1 पौलुसेरे तरफां ज़ै परमेशरेरो दास ते यीशु मसीहेरो प्रेरते, परमेशरेरे च़ूनोरे लोकां केरे विशवास, ते तैस सच़्च़ेरे पिशानानेरे मुताबिक ज़ै भेगतारे मुताबिके|
+\v 2 तैस हमेंशरी ज़िन्दगरे उमीदी पुड़, ज़ैसेरो वेइदो परमेशरे ज़ै झूठ न ज़ोए पेहिलो देंतेरो कीयोरोए|
+\v 3 पण ठीक वक्ते पुड़ अपनो वचन तैस प्रचारे सैहीं बांदो केरो, ज़ै इशे मुक्ति देंने बाले परमेशरेरे हुकमेरे मुताबिक मीं सोंफोरोए|
+\p
+\v 4 तीतुसेरे नोंवे ज़ैन विशवासे सैहीं मेरू सच़ू मठ्ठूए: बाजी परमेशर ते इशे मुक्ति देंने बाले प्रभु यीशु मसीहेरे तरफाँ अनुग्रह ते शान्ति भोती राए|
+\s क्रेते मां तीतुसेरू कम्म
+\p
+\v 5 मीं तू ऐलहेरे लेई करेते मां शारोरो थू, कि तू बाकी रेवरी गल्लन सुधारस, ते मेरे हुकमेरे मुताबिक नगरे नगरे मां बुज़ुरग च़ुनस|
+\v 6 ज़ैना बेइलज़ाम ते एकी एकी कौन्शरे मुन्श भोंन, ज़ैन केराँ मठ्ठाँ विश्वास केरतां भोंन, ज़ैन पुड़ एशपर्सती ते हुकम न मन्नेरो दोष न भोए|
+\v 7 किजोकी रखवालो परमेशरेरो भण्ढारी भोने सैहीं बे इलज़ाम लोड़े; न हठी, न सेरकलो, न पीने बालो, न मार पीट केरने बालो, ते न घटिया कुमैइयरो लालच़ी|
+\v 8 पण महैमानां केरी आदर केरने बालो, भलाई चहने बालो, बरदाश केरने बालो, अदालत केरने बालो, पवित्र ते झ़ैल्लने बालो भोए|
+\v 9 ते वफादार, वचनेरे मुताबिक बनोरो रह, कि रोड़ी तालीमी सैहीं शिखैली बटे; ते बेंस केरने बालाँ केरू आशी भी बंद केरी बटे|
+\p
+\v 10 किजोकी बड़े लोक बे फैइदी बेंस केरने ते धोखो देंने बालेन; ख़ास केरतां तैना ज़ैन केरो खतनो भोवरोए|
+\v 11 ऐन केरू आशी बंद केर्रू लोड़े: इना लोक नीच़ कुमैइयरे लेई बे पैइदी गल्लाँ शिखैलतां घरां केरां घरां बिगाड़तन|
+\v 12 तैने मैरां एके मैंन्हें ज़ै तैन केरो नबीए, ज़ोवरूए, कि क्रेती लोक सदा झूठे, दुष्ट जनवर ते आलसी पेटू भोतन|
+\v 13 ई गुवाही सच़ीए, ऐलहेरे लेई तैन खबरदार केर, कि तैना विशवासे मां पक्के भोन|
+\v 14 ते तैना युहूदी केरी कथा कहैन्न ते तैन मैंन्हू केरे हुकम्मन पुड़ मन न लान, ज़ैना सच़्च़े कैरां भटकतन|
+\v 15 शुद्ध लोकाँ केरे लेई सब चीज़ाँ शुद्धन, पण गंदे, ते विशवास न केरने बालाँ केरे लेई किछ भी शुद्ध नैइं: बल्कि तैन केरी अक्ल ते ज़मीर दुइये गंदेन|
+\v 16 तैना ज़ोतन, कि अस परमेशरे ज़ानतम: पण अपने कम्मन सैहीं तैसेरो इनकार केरतन, किजोकी तैना भिटे, ते हुकम न मन्ने बालेन: ते कोंची रोड़े कमेरे लायिक नैइं|
+\c 2
+\s आचरणेरे लेई रोड़ी तालीम
+\p
+\v 1 पण तू एरी गल्लां ज़ो, ज़ैना रोड़ी तालीमारे काबलन|
+\v 2 मतलब बुडडे मड़द, प्रहेज़ केरने बाले, ते समझ़दार ते संयमी भोंन, ते तैन केरो विशवास ते प्रेम, ते सबर पक्की भोए|
+\v 3 ऐंच़रे बुडडी कौन्शां केरो चाल चालन भी पवित्र लोकां केरो ज़ेरो भोए, दोष लाने बैली, ते पीने बैल्ली नैइं; पण रोड़ी गल्लाँ शिखालने बैली भोंन|
+\v 4 ताकि तैना जवान कौंन्शन खबरदार केरती रहन, कि अपने मुन्शन ते बच्चन सैहीं प्यार केरन|
+\v 5 ते बरदाश्त केरने बैली, पवित्र, घरेरो कारोबार केरने बैल्ली, भलाई केरने बैली ते अपने अपने मुन्शाँ केरे आधीन रहने बैली भोंन, ताकि परमेशरेरे वचनेरी निन्दा न भोए|
+\p
+\v 6 ऐंच़रे जुवान मड़दन भी समझ़ेइतां, कि संयमी भोंन|
+\v 7 सैभी गल्लन मां अपनो आप रोड़े कम्मन मां नमूनो बना: तेरी तालीमी मां सफाई, ते समझ़दारी भोए|
+\v 8 ते एरी खराई भोए, कि कोई तैस बूरी न ज़ोई बटे; ज़ैस सैहीं असन कोई भी दोष न हेरतां छिथोन|
+\p
+\v 9 गुलामन समझ़ा, कि अपने अपने मालकां केरे अधीन रहन; ते सैभी गल्लन मां तैन खुश रखन, ते तैन उलटो जुवाब न देन|
+\v 10 च़ोरी न केरन; पण सैभी किसमे सैहीं पूरे विशवासी निसन, कि तैना सैभी गल्लन मां इशे मुक्ति देंने बाले परमेशरेरी तालीमरी शोभा बधान|
+\p
+\v 11 किजोकी परमेशरेरो अनुग्रह बांदोए, ज़ै सैभी मैंन्हू केरे मुक्तरी वजाहे|
+\v 12 ते असन खबरदार केरते, कि अस अभेगती ते संसारेरे खुवैइशन मां मन बेदलतां इस युगे मां सयंम, ते धर्म ते भेगती सैहीं ज़िन्दगी बीताम|
+\v 13 ते बड़े उमीदी मतलब अपने महान परमेशर, ते मुक्ति देने बाले यीशु मसीहेरी महिमा बांदी भोंने बलगते रहम|
+\v 14 ज़ैने अपनो आप इशे लेई दितो, कि असन हर किसमेरे अधर्म कैरां छुटाए, ते खास केरतां अपने लेई अक एरी कौम बनाए ज़ै भला कम्मा केरे|
+\p
+\v 15 पूरे अधीकारे सैहीं इना गल्लाँ ज़ो, ते समझ़ा ते शिखालतो रा: कोई तीं धटया न समझ़े|
+\c 3
+\s मसीह चाल चलन
+\p
+\v 1 लोकन याद करा, कि हाकिमां, ते अधिकैरी केरे आधीन रहन, ते तैन केरो हुकम मन्न, ते हर एकी रोड़े कम्मेरे लेई तियार रहन|
+\v 2 कोंची बदनाम न केरन; लड़ाके न भोंन: पण नर्म सुभावेरे भोंन, ते सैभी मैंन्हन सैहीं साथी बेड़ी नेरमी सैहीं रहन|
+\v 3 किजोकी अस भी पेहिले, बे अकल, ते हुकम न मन्ने बाले, ते धोखो खाने बाले, ते रंगा रंगा केरी खूवैशन ते एशेरे गुलैमी मां थी, ते बैर, ते जलन रखने मां ज़िन्दगी बीताते थी, ते भिटे, ते एकी होरी सैहीं बैर रखते थी|
+\v 4 ज़ैखन इषो मुक्ति देंने बालेरी कृपा, ते मैंन्हन पुड़ तैसेरो प्रेम बांदो भोव|
+\v 5 त तैने असन मुखती दिती: ते इन रोड़े कम्मन सैहीं नैइं, ज़ैना असेईं एपू पुड़ अपने दियारे मुताबिक, नोंवें ज़रमेरे निहान, ते पवित्र आत्मा सैहीं असन नावें बनाने सैहीं|
+\v 6 ज़ै तैने इशे मुक्ति देंने बाले यीशु मसीहेरे ज़िरीये सैहीं असन पुड़ आढंगी उलटाई|
+\v 7 ज़ैस सैहीं अस तैसेरे अनुग्रह सैहीं धर्मी भोतम, हमेंशरी ज़िन्दगरे उमीदी सैहीं वारस भी बनम|
+\p
+\v 8 ई गल्ल सच़्च़े, ते आंव चहताइं, कि तू इना गल्लाँ केरे बारे मां दिलेरी सैहीं ज़ोस ऐलहेरे लेई ज़ैनेइं परमेशरेरू याकीन केरू, तैना भले भले कम्मन मां लगोरो रहनेरो ध्यान रखन: ते मैंन्हू केरे फैइदेरिन|
+\v 9 पण बेवकूफां केरी बेंसी, नसबनामा, ते बैर ते खलाफत, ते तैन लड़ाइन सैहीं, ज़ैना व्‍यवस्‍थारे बारे मां भोंन बांच़ोरो रा; किजोकी तैना बेफल ते बे फैइदेन|
+\v 10 कोंची पखेण्डी एकी दुइय बार समझ़ा ते अलग केर|
+\v 11 इन बुझ़तां कि एरो मैंन्हू टेपलोव, ते अपनो आप दोषी बनेइतां पाप केरतो रहते|
+\s खास हदायत ते अभिवादन
+\p
+\v 12 ज़ैखन आंव तुस्न अरतिमास या तुखिकुसे भेझ़ी, त मीं कां निकुपुलिस एजनेरी कोशीश केरां: किजोकी मीं तैड़ीए हिवतड़ो कटनो थापोरोए|
+\v 13 जेनास व्‍यवस्‍थापक ते अपुल्लोसे कोशीश केरतां पुज़ा, ते हेर, कि तैन कोंची चीज़री घटी न भोए|
+\v 14 ते इशे लोक भी ज़रूरतन पूरी केरनेरे लेई रोड़े कम्मन मां लगोरू रहनू शिखन ताकि बे फल न रहन|
+\p
+\v 15 मेरे सैभी सैथी केरू नमशकार, ते ज़ै विशवासेरे वजाही सैहीं असन सैहीं प्रेम केरतन, तैन नमसकार| तुस्न सैभन पुड़ अनुग्रह भोतो राए|
