+++ tests/fixtures/usfm/valid/id_tit_text_reg.usfm
@@ -1,0 +1,57 @@
+\id tit Regular
+\ide usfm
+\h Titus
+\toc1 Titus
+\toc2 Titus
+\toc3 tit
+\mt Titus
+\c 1
+Titus
+\v 1 Paulus, pelayan Allah dan rasul Yesus Kristus, untuk iman orang pilihan Allah dan pengetahuan akan kebenaran yang sesuai dengan kesalehan.
+\v 2 Semua ini dalam keyakinan akan hidup kekal bahwa Allah, yang tanpa kebohongan, menjanjikan sebelum segala jaman.
+\v 3 Pada saat yang tepat, Dia nyatakan firmanNya melalui pesan yang dipercayakanNya kepadaku untuk memberitakannya. Aku melakukan ini oleh perintah Tuhan Juruselamat kita.
+\v 4 Kepada Titus, anak yang sejati di dalam iman kita. Kasih karunia dan damai dari Allah Bapa dan Yesus Kristus Juruselamat kita.
+\v 5 Untuk tujuan ini aku meninggalkan engkau di Kreta, bahwa engkau mengatur segala sesuatu yang belum lengkap dan menempatkan penatua-penatua di setiap kota seperti yang ku perintahkan kepadamu.
+\v 6 Seorang penatua harus seorang tanpa cela, suami dari satu istri, dengan anak beriman tidak termasuk anak-anak jahat atau tidak disiplin.
+\v 7 Hal ini sangat penting bagi penilik, sebagai pengelola dari rumah tangga Allah, tidak bercacat cela. Dia tidak liar. Dia tidak cepat marah, tidak kecanduan anggur, bukan pembuat onar, dan bukan seorang yang tamak.
+\v 8 Sebaliknya, dia harus bermurah hati, seorang teman dari semua yang baik. Dia harus bijaksana, benar, saleh, dan dapat mengendalikan diri.
+\v 9 Dia harus memegang erat pesan yang dapat dipercaya yang telah diajarkan, dengan demikian dia mampu mendukung orang lain dengan ajaran baik dan mengoreksi orang-orang yang menentang dia.
+\v 10 Karena ada banyak orang yang menentang, secara khusus orang-orang bersunat. Perkataan mereka tidak berarti. Mereka menipu dan memimpin orang pada arah yang salah.
+\v 11 Kita harus menghentikan mereka. Karena mereka mengajar apa yang seharusnya tidak diajarkan untuk keuntungan yang memalukan dan memecah-belah seluruh keluarga.
+\v 12 Seorang dari mereka, orang bijaksana mereka, berkata, "Orang-orang Kreta suka berbohong, binatang-binatang yang jahat dan berbahaya, pelahap yang malas."
+\v 13 Pernyataan ini adalah benar, jadi koreksilah mereka agar mereka sehat di dalam iman.
+\v 14 Jangan memberi perhatian kepada mitos-mitos Yahudi atau kepada perintah-perintah orang yang menolak kebenaran.
+\v 15 Bagi orang yang murni, semuanya murni. Tetapi bagi orang-orang yang cemar dan tidak percaya, tidak ada yang murni. Karena pikiran dan hati nurani mereka cemar.
+\v 16 Mereka mengaku mengenal Allah, tetapi mereka menyangkal Dia dengan perbuatannya. Mereka menjijikkan dan tidak patuh. Mereka tidak diakui untuk setiap perbuatan baik.
+\c 2
+\v 1 Tetapi kamu, katakan apa yang sesuai dengan ajaran yang bisa dipercaya.
+\v 2 Kaum pria yang lanjut usia harus berkepala dingin, sopan, berpikiran sehat, teguh di dalam iman, di dalam kasih, dan dalam daya tahan.
+\v 3 Demikian juga kaum wanita lanjut usia harus menampilkan dirinya sebagai yang dihormati, tanpa gosip-gosip. Mereka harus tidak diperbudak oleh anggur. Mereka harus mengajarkan apa yang baik
+\v 4 dengan tujuan supaya melatih kaum wanita muda untuk mengasihi suami dan anaknya dengan tulus.
+\v 5 Mereka harus melatihnya menjadi bijaksana, murni, pengurus rumah tangga yang baik dan taat kepada suaminya. Mereka harus melakukan semua ini supaya dengan demikian perkataan Allah tidak dihina.
+\v 6 Demikian juga, mendorong kaum muda untuk menjadi bijaksana.
+\v 7 Di dalam segala hal tunjukkanlah dirimu sebagai teladan dari pekerjaan-pekerjaan baik; dan ketika engkau mengajar, tunjukkan integritas dan harga diri.
+\v 8 Sampaikanlah perkataan yang benar dan tidak tercela, sehingga siapapun yang menentang menjadi malu, karena dia tidak hal-hal yang jelek yang bisa dikatakan tentang kita.
+\v 9 Kaum budak harus taat kepada tuan-tuannya dalam segala hal. Mereka harus menyenangkan tuan-tuannya dan tidak membantah mereka.
+\v 10 Mereka harus tidak mencuri. Sebaliknya, mereka harus menunjukkan iman yang baik, sehingga dalam segala tindakannya mereka memuji ajaran kita tentang Allah Juruselamat kita.
+\v 11 Karena anugerah Allah telah menjadi nyata kepada semua orang.
+\v 12 Melatih kita untuk menyangkal kefasikan dan nafsu duniawi. Melatih kita untuk hidup bijaksana, adil, dan dalam cara hidup yang saleh di jaman ini
+\v 13 sambil kita mengantisipasi untuk menerima pengharapan yang penuh berkat, yaitu penyataan kemuliaan Allah yang besar dan Juruselamat kita Yesus Kristus.
+\v 14 Yesus telah memberikan diriNya bagi kita untuk membebaskan kita segala pelanggaran dan menyucikan, bagi diriNya, umat yang berharga yang berhasrat untuk melakukan pekerjaan yang baik.
+\v 15 Katakan dan anjurkan semua hal ini. Berikan koreksi dengan segala wibawa. Jangan biarkan seorangpun mengabaikan kamu.
+\c 3
+\v 1 Peringati mereka untuk tunduk kepada para penguasa dan pemerintah, taat kepada mereka, dan bersiap-sedia untuk melakukan setiap pekerjaan baik.
+\v 2 Peringati mereka untuk tidak mencaci-maki seorangpun, menghindari perdebatan, menghormati cara hidup orang lain, dan menunjukkan kerendahan hati kepada semua orang.
+\v 3 Karena kita dulu dungu dan tidak patuh. Kita dulu tersesat dan budak dari berbagai keinginan dan kesenangan. Kita hidup dalam kejahatan dan iri hati. Kita hidup dalam kejijikan dan saling membenci.
+\v 4 Tetapi kebaikan hati Allah Juruselamat kita dan kasihNya kepada umat manusia telah dinyatakan,
+\v 5 itu bukan karena pekerjaan kebaikan yang kita lakukan, melainkan oleh belas kasihanNya karena Dia telah menyelamatkan kita. Dia menyelamatkan kita melalui permandian dari kelahiran baru dan pembaharuan oleh Roh Kudus.
+\v 6 Allah mencurahkan Roh Kudus kepada kita melalui Juruselamat kita Yesus Kristus.
+\v 7 Dia melakukan ini supaya, kita yang dibenarkan oleh anugerahNya, sehingga menjadi yakin sebagai ahli waris dari kehidupan yang kekal.
+\v 8 Berita ini layak dipercaya. Aku ingin engkau memberitakan semua ini dengan penuh percaya diri, sehingga barang siapa yang percaya kepada Allah ingin melakukan pekerjaan baik yang telah diberikan olehNya kepada mereka. Semua hal ini baik dan menguntungkan bagi setiap orang.
+\v 9 Tetapi hindarilah perdebatan yang bodoh dan hal-hal mengenai silsilah serta perselisihan dan pertentangan mengenai hukum. Semuanya itu tidak berharga dan tidak menguntungkan.
+\v 10 Tolaklah setiap orang yang menyebabkan perpecahan diantara kalian setelah satu atau dua kali diperingatkan,
+\v 11 dan pahamilah bahwa ada orang-orang yang telah berbalik dari kebenaran dan berdosa serta menghukum dirinya sendiri.
+\v 12 Setelah saya mengutus Artemas atau Tikhikus kepadamu, segeralah datang kepadaku di Nikopolis dimana aku memutuskan untuk tinggal di sana selama musim dingin.
+\v 13 Segera utuslah Zenas, ahli Taurat itu, juga Apolos, bersama dengan segala yang mereka perlukan.
+\v 14 Orang-orang kita harus belajar untuk ikur serta dalam pekerjaan baik untuk memenuhi kebutuhannya sendiri supaya sampai mereka akan menghasilkan buah.
+\v 15 Semua orang yang bersamaku menyampaikan salam kepadamu. Salam kepada orang-orang yang mengasihi kita di dalam iman. Anugerah menyertai kamu sekalian.
