+++ tests/fixtures/usfm/valid/tc_commas.usfm
@@ -1,0 +1,51 @@
+\h Titus
+\id TIT, gux_Gourmanchéma_ltr, EN_ULB, Thu Jul 20 2017 16:03:48 GMT-0700 (PDT), tc
+\c 1
+\p
+\v 1 Mini polo , U tienu naaceeno, leni Jesu Kilisti tondo,ki baa gobidi U tienu n gandi yaaba li dandanli nni ke ban bandi ya moamoani ŋua U tienu tuoma. 
+\v 2 lan kuli tie ya miali kan gbeni yaa suginma tanpuuli. U tienu yua kan pua faama n den bili ya ŋoanianu hali mi cilima. 
+\v 3 O den bogini ke min doagidi O maama u yogunu n den pundi, nani U tienu ti faabikoa n den bili maama .
+\v 4 Fini tita, n bibuadiga, ti kuli dandanli po , U tienu ti Ba, leni Jesu Kilisti ti Candaano ŋanbili leni o yanduama n ya ye leni a. 
+\v 5 N den ŋa a kileta ki bua ŋan ŋanbi yaali n den sieni. nani min den teni a ya ñoabona.
+\v 6 Dogu n kuli ŋan kaani bi gobidikaaba, li kani ya pia ban tiedi yaaba yaa po o siedi ŋamo, ki tie puoyendo Calo, ke bi bila tie yaaba n tiega, yaaba n kaa pia cangindi, kaa tie biyenyieeda.
+\v 7 Li tie tiladi, ke li jaandieli gobidikoa n ya pia ya siedo n ŋani, nani U tienu naacemo yeni. wan da ya tie japaadaano, pafiima daano; yua n puni o yuli li dañoli. ka tie nibiado, ki da lingi ya baali n ki tiegi.
+\v 8 Ama ki gaani bi niba cangu, ki go tie bi niŋanba danli, yumiinli daano, ki tiega, ki gagidi ki ŋana, ki pia mi dingima,
+\v 9 yua n teni o yuli imoamoani maama, yaama ke o den bangima. Yeni ke wan fidi ki teni mi bangimoama, yaama n nia leni mi bangifaama.
+\v 10 Li pia bi  niyiekaaba boncianla, ama bi yaba yaaba n kua ku koanciagu siiga ki cie ki boandi bi yaba ki go maadi ma joagindi buoli kuli. 
+\v 11 Li tie tiladi ke landanba ñoana n mubini,   bi yadi ti diedi, ki tiendi mi bangima ki lingi ya ñuadi n pia yi fe, lan ki bi naani n ni ban bangi yaama.
+\v 12 Bi yaa sawalipuaba nni, yendo den yedi: kileta yaaba tie tuofadanba, muayanbiaditi,tugupialu danba.
+\v 13 li siedo tie moamoani. fu leni ba, ke ban ba li dandanmoanli .
+\v 15 Li bonla kuli ŋani yaaba n gagidi ki ŋani po, ama li baa kuli ki ŋani yaaba n juagini leni a pajuaga danba po, bi yanfuoma leni yantiana joagini. 
+\v 16 Bi tua ke bi bani U Tienu, ama ki nan nia leni O bi tuona nni, bi tie cangididanba, ni yiekaaba, ki kan fidi ki tieni tuonŋanli ba mo.
+\c 2
+\p
+\v 1 Ama fini wani han teni yaa bangima n tiegi.
+\v 2 Han yedi ke ajakpela n yaa tie yansoangidanba, leni moamoandanba, ki yaa se ke li hani, ki pia mi dingima leni mi gagidiŋanma li dandanli nni, leni mi buama nni,leni li juunli nni.
+\v 3 han go waani apociana moko ke ban yaa pia yaa tagi n naani leni mi gagidi hanma.ban da maadi mi ma biadima, ki da pa bi yula mi daama po, ke ban yaa tendi yaa tundi n hani,
+\v 4 ke ban fidi ki bangi bi pomadiba ban yaa bua bi ciaba leni bi bila,
+\v 5 ban tua yaaba n se ke li pa ki kubi bi diedi ke li hani,ki go pia li yumiinli bi ciaba kani,ke bi niba n da sugidi U Tienu maama.
+\v 6 Han go tundi bi nipanba ke ban yaa pia mi dingima,
+\v 7 fini mo n tua li bontogidikaala, a tuonhanma nni,ki yaa tendi yaa bangima nhani ki go tiegi, 
+\v 8 leni yaa maama n hani, ban kan fidi ki kuli yaama leni li ba, ke yibali n fiagi ki da ba wan ba maadi ti po ya mabiadima.
+\v 9 tundi bi tuonsoanba ban yaa miindi bi yula bi canbaanba kani,ki mangi bi pala libonla kuli nni, ki da yie ban yedi yaala,
+\v 10 ki da su liba, ama ki yaa wangi yogunu kuli yaa teginma n gagidi, ke ban teni bi niba n yaa jigindi U Tienu ti candaano bangima li bonla kuli nni.
+\v 11 kelima UTienu hanbili yaali n tie mi faabima hiali bi niba kuli po, den doagidi.
+\v 12 ki bangi ti ke tin yie yaala n kaa tiegi leni li handunli bon buakabiadila, yaa yogunu n tiena nni, leni mi yanfuoma, mi teginma leni U Tienu fangili,
+\v 13 ki yaa guu yaa suginma n pia li pamanli, U Tienu leni jesu kilisiti ti candaano doagidima po.
+\v 14 Wani yua n den puni o yuli ki daa ki faabi ti  leni ti tuonbiadi po ke tin tua wan die yaaba , o gagidi ti  ke tin yaa tuuni a tuonŋama kuli.
+\v 15 waani laa bonla ki fu leni ba ke o ba kuli n da fali a.
+\c 3
+\p
+\v 3 Kelima ti moko den kpa tie yanluodanba, niyiekaaba, niyaadikaaba leni li nunponli buoli kuli danba leni mi conconma danba ki ye li nunponbiadili nni, ti den tie lan pundi ban nani yaaba, ki go nani ti yaba.
+\v 4 Ama U Tienu ti candaano hanbili leni o buama n den doagidi bi niba kuli po 
+\v 5 o den faabi ti ke laa tie ti tuonhama bi ti teginma po ka. Ama o nihingbadima mi maaliema yaa batisima mi fuoma nni goa pandima po .
+\v 6 Wan den jiini tipo yaama ti candaano Jesu kilisiti nni.
+\v 7 Ke tinba yaaba n ba li moamoansaali kelima o ŋanbili yaa po, ki tua bi faajieba ki go ba yaa miali n kan gbeni yaa suginma.
+\v 8 Laa maama tie madugikaama. N bua ke han waani laa bonla kuli, ke yaaba n daani U Tienu n pa bi yula ki yaa tuuni a tuonhama kuli. Lani n hani ki go pia mi todima bi niba kuli po.
+\v 9 Ama han foagidi leni mi yanluomaama leni yaajanba togida, a koana leni yaa niama n tuugi leni li balimaama kelima li kuli tie fanma.
+\v 10 yaa nilo n fiini mi paadima han tundi o yenma ki pugini mi liema li yaa puoli han paadi leni o 
+\v 11 ki yaa bani ke laa nilo tieni ti tuonbiadi, ki yaadi, ki buni ki cuo o yuli.
+\v 12 N yaa ti soani apo Atenasa bi Tikika han jaligi ki cua n kani Nikapolisa kelima n jagi ke n baa ye lankane ku fawaagu yogunu.
+\v 13 Han bogini Sena li balimaama bangikoa leni Apolosa sancenli ke liba kuli n da poadiba.
+\v 14 Kali ti yaaba moko n bangi ki yaa tiendi yaala n hani ke ban fidi ki yaa baadi tipo yaala n tie tiladi, ke ban da tua yaa niba n kaa luoni bonluonkaala.
+\v 15 Yaaba n ye leni nni kuli fuondi a, han fuondi yaaba n bua ti li dandanli nni kuli. Ke li hanbili n yaa ye leni yi kuli.
