+++ tests/fixtures/usfm/valid/tc_spaces.usfm
@@ -1,0 +1,61 @@
+\id TIT EN_ULB sw_Kiswahili_ltr Wed Jul 26 2017 22:14:55 GMT-0700 (PDT) tc
+\h Titus
+\c 1
+\p
+\v 1 Paulo, mtumishi wa Mungu na mtume wa Yesu Kristo, kwa uaminifu wa Mungu alichagua watu na kwa maarifa ya kweli yaliyokubalika pamoja na utauwa. 
+\v 2 Walio na ujasiri katika uzima wa milele ambao Mungu kamwe hasemi uongo, uzima huu wa milele ulioahidiwa kabla ya nyakati zote. 
+\v 3 Kwa wakati sahihi, Mungu alilifunua neno lake kwa ujumbe kwamba aliniamini mimi kuuhubiri. Nilifanya hivi kwa amri ya Mungu, mwokozi wetu.
+\v 4 Kwa Tito, mwana wa kweli katika imani yetu. Neema, huruma na amani kutoka kwa Mungu Baba na Yesu Kristo mwokozi wetu. 
+\v 5 Kwa sababu hii nalikuacha Krete, kwamba uyatengeneze mambo yote yaliyokuwa hayajakamilika na kuweka wazee wa kanisa kwa kila mji kama nilivyokuagiza.
+ <<<<<<< HEAD:mergetest
+\v 6 Mzee wa kanisa lazima asiwe na lawama, mme wa mke mmoja, aliye na watoto waaminifu wasioshuhudiwa mabaya au wasio na nidhamu. 
+\v 7 Ni muhimu kwa mwangalizi, kama msimamizi wa nyumba ya Mungu, asiwe na lawama. Asiwe mtu wa kelele au asiye jizuia. Lazima asiwe mwepesi wa hasira, asiwe mlevi, asiwe mtu wa kusabisha ugomvi, na asiwe mwenye tamaa.
+\v 8 Badala yake: awe mtu mkaribishaji, mwenye wema. Lazima awe mtu mwenye akili timamu, mwenye haki, mcha Mungu, anayejitawala mwenyewe. 
+\v 9 Awezaye kusimamia mafundisho ya kweli, anayeweza kuwatia moyo kwa mafundisho mazuri na awezaye kuwarekebisha wote wampingao.
+ =======
+\v 6 Mzee wa kanisa lazima asiwe na laasfasa fasn fdnas dye na watabaya au wasio na nidhamu. 
+\v 7 Ni muhasdfmlads as fjkasgalizi, kama ma. Asiwe mtu wa kelele au asiye jizuia. Lazima asiwe mwepesi wa hasira, asiwe mlevi, asiwe mtu wa kusabisha ugomvi, na asiwe mwenye tamaa.
+\v 8 Badaawe mtu mwenye akili timamu, mwenye haki, mcha Mungd kmlskfmas   itawala mwenyewe. 
+\v 9 Awezaye kusimamia mafundisho yaasfasdfasdisho mazuri na awezaye kuwarekebisha wote wampingao.
+ >>>>>>> 4e2b407f501b68f8588aa645acafffa0224b9b78:mergetest
+\v 10 Kwa kuwa kuna waasi wengi, hasa wale wa tohara. Maneno yao ni ya upuuzi, wanadanganya na kuongoza watu katika upotovu. 
+\v 11 Ni lazima kuwazuia watu kama hao. Wanafundisha yale wasiyopaswa, aibu kwa faida na kuharibu nyumba nzima.
+\v 12 Mmoja wao, mtu wa busara, alisema, "Wakrete wana uongo usio na mwisho, wabaya na wanyama wa hatari, wavivu na walafi." 
+\v 13 Haya maelezo ni kweli,kwa hivyo uwazuie kwa nguvu ili kwamba waweze  kusema kweli katika imani.
+\v 14 Wewe usijihusishe na hadithi zisizo za kweli  za Kiyahudi au amri za watu, ambao hurudisha nyuma ukweli.
+\v 15 Kwa wote walio safi, kila kitu ni safi. Lakini kwa wote walio wachafu na wasioamini, hakuna kilicho safi. Kwa kuwa mawazo yao na dhamira zao zimechafuliwa. 
+\v 16 Wanakiri kumjua Mungu, lakini kwa matendo yao wanamkana. Wao ni waovu na wasiotii na hawakuthibitishwa kwa tendo lolote jema.
+\c 2
+\p
+\v 1 Lakini wewe yaseme yale yanayoendana na maelekezo ya kuaminika. 
+\v 2 Wazee wawe na kiasi, heshima, busara, imani safi, katika upendo, na katika uvumilivu.
+\v 3 Vilevile wanawake wazee lazima daima wajionyeshe wao wenyewe kama wenye kujiheshimu, na sio wasengenyaji. Lazima wasiwe watumwa wa pombe. 
+\v 4 Wanapaswa kufundisha yaliyo mazuri ili kuwaandaa wasichana kwa busara kuwapenda waume zao na watoto wao. TEST 
+\v 5 Wanapaswa kuwafundisha kuwa na busara, wasafi, watunzaji wazuri wa nyumba zao, na watii kwa waume zao wenyewe. Inawalazimu kufanya mambo haya ili kwamba neno la Mungu lisitukanwe.
+\v 6 
+\v 7 Katika njia zote jiwekeni wenyewe kuwa mfano katika kazi nzuri; na mfundishapo, onesheni uadilifu na heshima. Ongeeni ujumbe ambao ni wenye afya na usio na dosari, ili kwamba yeyote apingaye aaibishwe kwa sababu hana baya la kusema juu yetu.
+\v 8 katika namna hiyo hiyo, watieni moyo vijana wa kiume wawe na busara. 
+\v 9 Watumwa na wawatii mabwana zao katika kila kitu. Wanapaswa kuwafurahisha na sio kubishana nao. 
+\v 10 Hawapaswi kuiba. Badala yake, wanapaswa kuonesha imani yote nzuri ili kwamba katika njia zote wayapambe mafundisho yetu kumhusu Mungu mwokozi wetu.
+\v 11 Kwa kuwa neema ya Mungu imeonekana kwa watu wote. 
+\v 12 Inatufundisha kukataa mambo yasiyo ya kimungu na shauku za kidunia. Inatufundisha kuishi kwa busara, kwa haki, na katika njia za kimungu kwa wakati huu. 
+\v 13 Wakati tunatarajia kupokea tumaini letu lenye baraka, mwonekano wa utukufu wa Mungu wetu mkuu na Mwokozi wetu Yesu Kristo.
+\v 14 Yesu alijitoa mwenyewe kwa ajili yetu ili kutukomboa kutoka katika uasi na kutufanya wasafi, kwa ajili yake, watu maalumu walio na hamu ya kufanya kazi nzuri.
+\v 15 Yaseme na kuyasisitiza mambo haya. Kemea kwa mamlaka yote. Usikubali mtu yeyote akupuuze.
+\c 3
+\p
+\v 1 Uwakumbushe kuwanyenyekea viongozi na mamlaka, kuwatii na kuwa tayari katika kila kazi njema. 
+\v 2 Wakumbushe kutokumtendea mtu yeyote ubaya, waepuke majibizano, wape nafasi watu kuamua mambo yao, na kuonyesha unyenyekevu kwa watu wote.
+\v 3 Kwa kuwa sisi pia tulikuwa na mawazo potofu na wakaidi. Tuliokuwa upotevuni na kufanywa watumwa kwa tamaa nyingi na starehe. Tuliishi katika wivu na uovu. Tulichukiza na kuchukiana.
+\v 4 Lakini huruma ya Mungu mwokozi wetu na upendo wake kwa wanadamu ulipoonekana, 
+\v 5 haikuwa kwa matendo yetu ya haki tuliyofanya, bali alituokoa kwa neema yake. Alituokoa kwa kutuosha kwa kuzaliwa upya na katika kufanywa upya kwa Roho Mtakatifu.
+\v 6 
+\v 7 Mungu alimwaga Roho Mtakatifukwa wingi  juu yetu kupitia Yesu Kristo mwokozi wetu. Alifanya hivi ili kwamba, tukiwa tumehesabiwa haki kwa neema yake, tuwe washirika katika uhakika wa maisha ya milele.
+\v 8 Huu ni ujumbe wa kuaminika. Ninawataka myanene kwa ujasiri mambo haya , ili kwamba wale wanaomwamini Mungu wawe na dhamira juu ya kazi nzuri ambayo aliiweka mbele yao. Mambo haya ni mazuri na yanafaida kwa ajili ya watu wote. 
+\v 9 Lakini jiepusheni na mijadala ya kipumbavu, nasaba, mashindano, na migongano kuhusu sheria. Mambo hayo hayana maana wala faida. 
+\v 10 Mkataeni yeyote anayesababisha migawanyiko kati yenu. Baada ya onyo moja ama mawili, 
+\v 11 jueni kwamba mtu wa jinsi hiyo ameiacha njia iliyo sahihi na anatenda dhambi na kujihukumu mwenyewe.
+\v 12 Nitakapomtuma kwako Artemi au Tikiko, harakisha uje kwangu huku Nikopoli ambapo nimeamua kukaa kipindi cha baridi. 
+\v 13 Harakisha na umtume Zena mtaalamu katika sheria, na Apolo, pasipo kupungukiwa na kitu.
+\v 14 Watu wetu lazima wajifunze kujishughulisha katika kazi nzuri ambazo zinagusa mahitaji muhimu ili  kwamba wasije wakawa wasiozaa matunda.
+\v 15 Wote walio pamoja nami wanakusalimu. Wasalimu wale watupendao katika imani. Neema iwe nanyi nyote.
