+++ tests/g.GitApi.js
@@ -16,7 +16,7 @@ describe('GitApi.status', function() {
     })
   });
   it ('status should give an error when not in a git repo', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     fs.removeSync('../testDir');
     fs.removeSync('../testRepo');
     fs.ensureDirSync('../testDir');
@@ -28,7 +28,7 @@ describe('GitApi.status', function() {
     })
   });
   it('status should give the status of the current repo', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi().status(function(err, data) {
       assert.isNull(err);
       assert.isObject(data);
@@ -40,14 +40,14 @@ describe('GitApi.status', function() {
 
 describe('GitApi.checkout', function() {
   it ('checkout should give an error when not in an existing directory ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').checkout('test', function(err) {
       assert.isNotNull(err);
       done();
     })
   });
   it ('checkout should give an error when no branch is given ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').checkout(null, function(err) {
       assert.isNotNull(err);
       assert.equal(err, "No branch");
@@ -56,7 +56,7 @@ describe('GitApi.checkout', function() {
   });
 
   it ('checkout should give an error when not in a git repo', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../testDir').checkout('master', function(err) {
       assert.isNotNull(err);
       done();
@@ -66,14 +66,14 @@ describe('GitApi.checkout', function() {
 
 describe('GitApi.mirror', function() {
   it ('mirror should give an error when not in an existing directory ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').mirror('test', '../invalidFolder', function(err) {
       assert.isNotNull(err);
       done();
     })
   });
   it ('mirror should give an error when no data is given ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').mirror(null, null, function(err) {
       assert.isNotNull(err);
       assert.equal(err, "Missing URL or save path");
@@ -82,8 +82,8 @@ describe('GitApi.mirror', function() {
   });
 
   it ('mirror should clone a git repo into a directory', function(done) {
-    this.timeout(50000)
-    GitApi('../testDir').mirror('https://git.door43.org/royalsix/test1133', '../testRepo', function(err) {
+    this.timeout(2000)
+    GitApi('../testDir').mirror('https://git.door43.org/klappy/blank.git', '../testRepo', function(err) {
       assert.isNull(err);
       done();
     })
@@ -92,14 +92,14 @@ describe('GitApi.mirror', function() {
 
 describe('GitApi.add', function() {
   it ('add should give an error when not in an existing directory ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').add(function(err) {
       assert.isNotNull(err);
       done();
     });
   });
   it ('add should give an error when not in a git repo', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     fs.ensureDirSync('../testDir');
     GitApi('../testDir').add(function(err) {
       assert.isNotNull(err);
@@ -107,7 +107,7 @@ describe('GitApi.add', function() {
     })
   });
   it('add should add files of the current repo', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi().add(function(err) {
       assert.isNull(err);
       done();
@@ -117,14 +117,14 @@ describe('GitApi.add', function() {
 
 describe('GitApi.init', function() {
   it ('init should give an error when not in an existing directory ', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../invalidFolder').init(function(err) {
       assert.isNotNull(err);
       done();
     });
   });
   it ('init should initialize a git repo with no issue', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     fs.ensureDirSync('../testDir');
     GitApi('../testDir').init(function(err) {
       assert.isNull(err);
@@ -135,7 +135,7 @@ describe('GitApi.init', function() {
 
 describe('GitApi.add', function() {
   it ('add should add an untracked file to staging', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../testDir').add(function(err) {
       GitApi('../testDir').status(function(err, data) {
         assert.isNull(err);
@@ -148,7 +148,7 @@ describe('GitApi.add', function() {
 
 describe('GitApi.commit', function() {
   it ('commit should commit staged files', function(done) {
-    this.timeout(500000);
+    this.timeout(1000);
     GitApi('../testDir').commit('user', 'test commit', function(err, data) {
       assert.isNull(err);
       assert.isObject(data);
@@ -163,7 +163,7 @@ describe('GitApi.commit', function() {
 
 describe('GitApi.checkout', function() {
   it ('checkout should checkout to master', function(done) {
-    this.timeout(50000)
+    this.timeout(1000)
     GitApi('../testDir').checkout('master', function(err, data) {
       assert.isNull(err);
       GitApi('../testDir').status(function(err, data) {
